Fix #124167: Crash opening file after asset browser is closed
Switching to an asset browser, then back to a differend editor would
cause a crash when loading a new file.

Basically the file/asset browser tried to free dangling asset and asset
library pointers when trying to free itself. That's because the asset
system gets destructed with a `AS_asset_libraries_exit()` call before
the screen and with that the asset browser were freed.

More context:
- The file list has pointers to assets owned by the asset system
- The pointers are non-owning, however the file-list manages loading and
  unloading of the assets into/from the asset libraries.
- This creates a tricky to manage lifetime dependency between the file
  list and the asset library storage.

In C++, `std::weak_ptr` can be used for such circular dependencies. The
owner uses a shared, the non-owner uses a weak pointer. Whenever the
non-owner accesses the object, it first checks if the object is still
alive through the weak pointer interface.

Unfortunately this adds some noice in the changes:
- Assets need to be stored as shared pointers
- `FileListInternEntry` needs to be non-trivial to hold weak pointers.
  I added explicit member initializers for all members to make sure the
  result matches the previous `MEM_cnew()` call still, to play safe.