Node Tools: Add Random Seed node
This node returns a pseudo-random seed every time it is executed.

The idea is that the node can be used for convenience when the
users wants a `Seed` socket to be random for every execution,
instead of having to tweak the seed manually every time.

Internally, the node uses the current time in milliseconds to generate
the hash. The resolution could also be increased to microseconds if
we think milliseconds is not enough.