Fix: GPv3: Duplicate previous key using auto key
When erasing, sculpting or tinting, the previous key
was not being duplicated.

The `ed::greasepencil::ensure_active_keyframe`
didn't have a good mechanism to make this work.
This adds a parameter to the `ensure_active_keyframe`
function so that the caller can decide what should
be done.

For the sculpting tools, eraser, and tint tool, this
will now duplicate the previous key, when auto-key
is on.