Fix #125219: Transmissive BSDF different in Blended and Dithered mode
for transmissive closures the color is squared, but Blended square
BEFORE the weight is applied, and Dithered square AFTER the weight is
applied, causing difference in the final color.

This fix aligns the behaviour of Blended with Dithered.

NOTE that this is a partial fix, because Dithered is also behaving
incorrectly. A proper fix would be !125233.