Readfile: Improve memleaks info in debug builds.
Refactor the storage of 'alloc names' for blocks read from blendfiles,
by storing them in a util static data. Compared to previous code, this
change:
- Fixes ASAN aborting before readfile-related memleaks prints.
- In addition to the 'main owner' data type, add the actual read struct
  type (and array size if relevant) to the alloc name.

Note that this keeps all combinations of owner/types/array_num generated
strings for the whole duration of the program. However, this only
impacts Debug (and Release `WITH_ASSERT_ABORT`) builds, not 'real'
release ones, so this seems worth the extra information.

To avoid a global static variable, the "construct on first use" idiom is
used. However, since that storage needs to be initialized before the
MEM_guardedalloc leak detector, it requires exposing a
`BLO_init_readfile_alloc_info_storage()` utils, that needs to be called
from the very early initialization code in `main` function of
`creator.cc`.