Anim: UI template for Action selectors
Add a UI template `UILayout.template_action()`. It allows selecting the
Action that is assigned to a data-block.

Normally `template_ID()` would be used for such things, but that is not
usable for assigning Actions. The issue is that that template requires
an RNA property to exist to operate on. Such a call (from Python) would
look like `layout.template_ID(id.animation_data, "action", ...)`, but
when the ID has never been animated, `id.animation_data` is `None`,
which has no "action" property.

The `template_action()` template is coded to deal with these cases.