Fix #125353: Select by active material fails to select the selected slot
Caused by 296d05060d22

This was the case if multiple slots had the same material, then the
first corresponding slot was taken (instead of the preferred active).

Now restore the special behavior from 6b39dc7672eb but without the issue
of clamping the slot index (which was the problem in #123560). This is
done by introducing a non-clamping version of `BKE_object_material_get`.