Refactor: Anim, use enum types for enum values in bAnimContext
Use actual enum types (`eAnimEdit_Context` etc) for enum values in the
`bAnimContext` struct.

This commit also splits the `mode` field into `dopesheet_mode` and
`grapheditor_mode`, so that their values are clearly differentiated.
Before you'd have to check `spacetype` to know how to interpret this
field (but that wasn't described either). You still have to check that
field to know which one of the two you need to look at, but at least
it's now clear what values can be expected there.