Compositor: Add shared derived resources optimization
This patch adds a new Derived Resources concept to the GPU compositor.
Derives resources are resources that are computed from a particular
result and cached in it in case it is needed by another operation, which
can greatly improve performance in some cases at the cost of more memory
usage.

This is a prerequisite for the new canvas/transform handling that is
planned for the compositor. But it is currently only implemented for the
denoise operation as a prototype.

This makes multi-pass compositing setups two times faster.