Fix #125680: Dedicated overlay shader for edit GPv3
Grease Pencil v3 was using the same shader as particle strands, this was
meant to be temporary during the development. Now two separate dedicated
shaders are created for GPv3 use only and whey use grease pencil colors
and point sizes from the user preferences instead of resuing those of
meshes. This gives us a basis for implementing different overlay drawing
for GPv3 in the future as well.