Refactor how tools handle brush info
There's an option now to flag tools as using brushes. Previously this was done
by setting the `data_block` field for tools, but this was weird naming, and was
causing issues with the logic to switch brush assets when switching tools.

Dynamically setting `data_block` based on the current brush, as we'd do with
brush assets now, means that we couldn't differentiate anymore between tools
that supported any brush, or only brushes of a certain type. The `data_block`
field is still there, but now is used only to indicate that a tool requires a
certain brush type. I can rename this later still.

This way we know now if a tool requires a certain brush type or supports any
brush type, and can set the active brush accordingly when switching tools.