Fix: Cycles BVH2 not computing N closest intersections
the code snippet is supposed to compute the maximal `isect.t` in the
array, which is used to determine if subsequent intersections should be
added.
However, the previous implementation includes the old `isect.t` which is
going to be replaced, resulting an overestimation of `tmax_hits` and
thus missing closer intersections.
This commit fixes the issue by computing the maximal `isect.t` after new
entry is inserted.