Fix #125558: Show UI_BTYPE_CURVE Control Points
The complaint of curve control points being almost invisible is because
of a missing call to GPU_program_point_size. The related complaint that
these were "very small" previously is because of a small maximum size.
This PR adds the missing call to allow the setting dot size, then
increases the minimum, regular, and maximum sizes.