Fix #125669: Also update `func_ArgN` functions from new button
When updating the old active button from the new layout `func_ArgN` was
swapped over but not the corresponding free and copy functions. This
could lead to `func_ArgN_free_fn` being null despite `func_ArgN` having
a value causing a crash when freeing the button.

---

While this fixes the described crash, there can now be a visible glitch,
where the separator between the buttons to the right of the ui list
becomes highlighted for a split second while the layout is updated.
This makes me suspect that the issue is partially that
`ui_but_equals_old`, the comparison function that matches old and new
buttons when rebuilding the layout, isn't doing a great job here,
similarly to #114461.