UI: Add LTS to Version String for Title Bar (4.2)
This PR makes the same changes to 4.2LTS as we have in main for the
formatting of the version string for use on the title bar and the
status bar (#125332 minus the alteration in #125788).  Title bar shows
a more detailed version string, while status bar shows more compactly.
"LTS" is included in the long form when defined. Patch version shown
for both. "Alpha", "Beta", "Release Candidate" included in long form,
but uses " a", " b", " RC" for short form.