Cleanup: Cycles: Consolidate coordinate system conversions
- Deduplicate Fisheye projection code
- Replace spherical/cartesian conversions with shared helpers
- Replace transforms from/to local coordinate systems with shared helpers

The main type of repeated transform that's not covered here is to/from_coords,
but with separate values for xy and z (e.g. BSDFs that already computed
dot(wi, N) earlier, so they only need dot(wi, X) and dot(wi, Y) later).
Could also be replaced, but it would feel weirdly specific for a helper
function. IDK.