Fix #117367: OBJ: fix vertex colors imported scrambled up
Lift certain incorrect assumptions about the order of vertices in
an OBJ file when processing vertex colors, which could lead to
some verts getting other verts colors.

Replaces the list-of-blocks representation, attuned to assumptions
that the verts in an object form a contiguous subrange, with a
flat array better suited for random-access.