Fix #126252: Polyline target circle changes color randomly
When the operator was introduced, the program specified the immediate
mode program GPU_SHADER_3D_UNIFORM_COLOR but did not bind a particular
value. This leads to the color used for this circle changing depending
on the last color bound when drawing.

This change makes the color of the circle white, as initially intended,
and adds a darker outline to improve readability.