Fix 124565: EEVEE Environment Render Pass
There were a couple of regressions when using the environment render
pass.

- Not working when film transparency was enabled
- Deferred geometry didn't occlude the background

This PR splits the world background shader (eevee_surf_world) into two
shaders. One clears all the render passes and a second one that renders
the background and the environment.

The second shader is moved after the deferred pipeline and an early
depth test is used to occlude the background and environment pass.

NOTE: render test reference images needs to be updated.