Fix #126149: Flat face wireframe edges hidden on some platforms
d1049f60821502f3d612 translated the logic from the `int8` to `float`
conversion somewhat incorrectly. Previously the data was extracted to
an integer array (multiplied by 254) first then translated to a float
VBO (and divided by 255). I think conceptually the special `1.0/255`
value isn't necessary because the "optimal display edges" area already
taken into account for the creation of the index buffer. But as a quick
fix, just fix the factor to be the same as before.