Removed the duplicated_frames_buffer from GPv3 transform data.
The purpose of this buffer was to keep an explicit list of frames that
are removed again after the operator is cancelled. This can be a simple
flag. When using `Map` these frames have to be kept in separate
containers to avoid overwriting each other during transform which would
destroy data. Now frame data will be stored in a simple array, making
separate lists unnecessary.