Nodes: Expose additional inputs
This takes some aspects of a previous Gabor patch (#110802)
and new inputs to the new implementation of Gabor noise.

The rationale is to provide artists with more control over the
Gabor noise and its performance.

New Inputs

Radius - This controls the radius of the kernel size. At small sizes
the kernel is smaller and covers less space.

Phase Shift - Shifts the phase of the Gabor noise as it is based
on sinusoidal functions.

Impulses - Controls the density of impulses per cell. At small values
coverage decreases but performance increases and vice versa.

The patch also adds a Normalize option similar to Voronoi and Noise.