UI Test: Allow Changing Editor Outline Width
This is not a proposal for a feature. It is a tool to help evaluate
differing editor outline widths. Once applied you can change the widths
in Preferences / Interface / Display.  It has been mentioned in a
meeting that a change in this width might be considered. This does not
consider line width, only scale (on purpose). It does not alter the
"hit" size for resizing but that would be affected by an actual change
so can be imagined.